+++ src/main/java/org/basex/query/value/type/FuncType.java
@@ -189,7 +189,7 @@ public class FuncType implements Type {
   }
 
   @Override
-  public int id() {
+  public byte id() {
     return 7;
   }
